requirements.txt
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==2.0.0
adafruit-circuitpython-framebuf==1.6.5
adafruit-circuitpython-requests==3.2.8
adafruit-circuitpython-ssd1306==2.12.17
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.63.0
Adafruit-PureIO==1.1.11
bleak==0.21.1
numpy==1.26.4
pigpio==1.78
pillow==10.3.0
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
setuptools==69.5.1
spherov2==0.12.1
transforms3d==0.4.1
typing_extensions==4.11.0
